Move the app state out of the main view
This allows the state to be an ObservableObject passed between
views, which allows the main App to respond to events which
only it can, like menu interactions, and cause the UI to
change in response to those.

It's also just cleaner to not have all of this in the view.

This change wasn't exhaustively tested, and it was a pretty
big move of a lot of code, so there's a chance that it broke
something, making it a good place to look if a bug is found.